update branch (#4)
* TCM 218943 weather station: support negative temperature

* generic_dimmer: Fix inconsistency

The default max values were defined to different values in the checkValues and
createCode functions, causing the webgui to not transmit slider changes to
values above 10.
This commit fixes this by setting the default to 15 in both functions using a
global define.

* Webgui: Fix dimmer display bugs

This would cause the text field next to the slider to be
initialized to 0 at page load, even though the slider was showing the
correct dimlevel.

This commit fixes this by also assigning the dimlevel to the text field on
page load. Patch by @Niek.

Further, the slider-input field was editable, but without use (editing the
value has no effect). Therefore, display the field as disabled.